sbc documents loading updated
