+++ db/seedfiles/plan_xmls/me/sbcs/2023/2023_sbcs.csv
@@ -1,0 +1,26 @@
+hios_id, sbc, year
+33653ME0262304-01,33653ME026230401-0922.pdf,2023
+33653ME0262305-01,33653ME026230501-0922.pdf,2023
+33653ME0262306-01,33653ME026230601-0922.pdf,2023
+33653ME0262308-01,33653ME026230801-0922.pdf,2023
+33653ME0262309-01,33653ME026230901-0922.pdf,2023
+33653ME0262310-01,33653ME026231001-0922.pdf,2023
+33653ME0262312-01,33653ME026231201-0922.pdf,2023
+33653ME0262313-01,33653ME026231301-0922.pdf,2023
+33653ME0262315-01,33653ME026231501-0922.pdf,2023
+33653ME0262316-01,33653ME026231601-0922.pdf,2023
+33653ME0272305-01,33653ME027230501-0922.pdf,2023
+33653ME0272306-01,33653ME027230601-0922.pdf,2023
+33653ME0272307-01,33653ME027230701-0922.pdf,2023
+33653ME0302301-01,33653ME030230101-0922.pdf,2023
+33653ME0302303-01,33653ME030230301-0922.pdf,2023
+33653ME0302304-01,33653ME030230401-0922.pdf,2023
+33653ME0632305-01,33653ME063230501-0922.pdf,2023
+33653ME0642301-01,33653ME064230101-0922.pdf,2023
+33653ME0642302-01,33653ME064230201-0922.pdf,2023
+33653ME0642303-01,33653ME064230301-0922.pdf,2023
+33653ME0642304-01,33653ME064230401-0922.pdf,2023
+33653ME0642305-01,33653ME064230501-0922.pdf,2023
+33653ME0652301-01,33653ME065230101-0922.pdf,2023
+33653ME0652302-01,33653ME065230201-0922.pdf,2023
+33653ME0652303-01,33653ME065230301-0922.pdf,2023
